+++ common/src/modules/gossip_messages.rs
@@ -1,0 +1,4 @@
+/// The gossip_messages module exchanges messages with the 
+/// other connected nodes.
+pub struct Module {
+}
